Construct AltosMapInterface
Provide an interface between the MapFragment and the map presentation itself.

+++ app/src/main/java/org/altusmetrum/altosdroid/TelemetryState.java
@@ -59,7 +59,7 @@ public class TelemetryState {
     }
 
     public void remove(int serial) {
-        states.remove((Integer) serial);
+        states.remove(serial);
     }
 
     public Set<Integer> keySet() {
